# Speed up core test