Resolves L2B-10355

Run | Before | After
-- | -- | --
1 | 639ms | 368ms
2 | 616ms | 259ms
3 | 587ms | 261ms
4 | 583ms | 263ms
5 | 585ms | 265ms
AVG | 602ms | 283.2ms

On average it's around 212% faster. So our CI shouldn't be timing out for some time